requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.5.3
pydantic-settings==2.1.0
motor==3.6.0
pymongo==4.9.2
redis==5.3.0
celery==5.4.0
requests==2.32.3
python-multipart==0.0.9
python-dotenv==1.0.1
cloudinary==1.43.0
pillow==11.0.0
opencv-python==4.10.0.84
pytesseract==0.3.10
paddleocr==3.4.1
paddlepaddle==3.3.1
transformers==4.41.0
torch==2.11.0
numpy==2.3.4
scikit-learn==1.6.0
pandas==2.2.3
httpx==0.27.0
reportlab==4.2.2
